Include time steppers module before diagnostics and output writers
