+++ src/TimeSteppers/TimeSteppers.jl
@@ -22,13 +22,17 @@ using Oceananigans.SurfaceWaves
 using Oceananigans.BoundaryConditions
 using Oceananigans.Solvers
 using Oceananigans.Models
-using Oceananigans.Diagnostics
-using Oceananigans.OutputWriters
 using Oceananigans.Utils
 
-using Oceananigans: AbstractTimeStepper
 using Oceananigans.TurbulenceClosures:
     calculate_diffusivities!, ∂ⱼ_2ν_Σ₁ⱼ, ∂ⱼ_2ν_Σ₂ⱼ, ∂ⱼ_2ν_Σ₃ⱼ, ∇_κ_∇c
+
+"""
+    AbstractTimeStepper
+
+Abstract supertype for time steppers.
+"""
+abstract type AbstractTimeStepper end
 
 """
     TimeStepper(name, args...)
